fixed app name, added npm start to run server and webpack, fixed location of database index
