+++ package.json
@@ -8,7 +8,7 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "build": "node setup.js",
+    "start": "nodemon server/index.js | webpack -d --watch",
     "test": "jest"
   },
   "dependencies": {
