+++ public/index.html
@@ -3,11 +3,11 @@
 
 <head>
   <title>Info</title>
-  <link rel="stylesheet" href="styles.css" />
+  <link rel="stylesheet" href="info.css" />
 </head>
 
 <body>
-  <div id="app">
+  <div id="info">
   </div>
   <script src="bundle.js"></script>
 </body>
